// src/utils/types/itemManageType.ts
export interface IItemMageTypeProps {
  makersId: number;
  calorie: number | undefined;
  carbohydrate: number | undefined;
  fat: number | undefined;
  itemIntro: string | undefined;
  itemName: string;
  price: number | undefined;
  protein: number | undefined;
  supplyPrice: number | undefined;
}